# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "graphrag",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Create Document",
        "description": "Multipart upload. Requires `Idempotency-Key`. Validates MIME by CONTENT SNIFFING (never\nthe client-declared `Content-Type`), plus a size cap. Returns 202 `{job_id, doc_id,\ncorrelation_id}` for a new document, or 200 with the SAME shape (no new job) when the\ncontent (by sha256) is already registered \u2014 content-addressing is what makes a re-upload\nunder a different Idempotency-Key still resolve to the original doc_id.",
        "operationId": "create_document_v1_documents_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentJobResponse"
                }
              }
            }
          },
          "200": {
            "description": "Document already exists (duplicate sha256).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{doc_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "description": "Enqueues `DeleteDocument`. Returns 202. Never deletes inline \u2014 see\n`services.ingestion.service.DeletionService`.",
        "operationId": "delete_document_v1_documents__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_document_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_document_v1_documents_post"
      },
...
      "DocumentJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "doc_id",
          "correlation_id"
        ],
        "title": "DocumentJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}